Solid Textures

+++ RayTracing/RayTracing/camera.h
@@ -24,7 +24,7 @@ public:
 		time0 = t0;
 		time1 = t1;
 		lens_radius = aperture / 2;
-		float theta = vfov * M_PI / 180;
+		float theta = (float)(vfov * M_PI / 180.0f);
 		float half_height = tan(theta / 2);
 		float half_width = aspect * half_height;
 		origin = lookfrom;
